[cli] Add option to avoid failing on invalid arg.
The previous behavior of `ParseResults` is to raise when requesting the
value of an argument that was not defined. This is a good thing in
general, as inspecting value of an option that was not specified is
often a sign of a bug. In certain instances, however, it is desirable
not to raise and return `None()` instead for `maybeGetString()`
etc. It is the case when a helper function builds an object from
`ParseResults` (that may originate from different `Command`
definitions) with default values for undefined arguments.