Do not rebuild the author of a privacy change
When any privacy change affecting a user is detected, we rebuild all
of their mirrored tables.

This change only does this if the user was not the author of the
privacy change. This is useful because:

1. first and least, they know about the change so it's very
conservative to rebuild all of their tables. It is expensive, a bad
experience on slow networks as the user can observer the intermittent
state, and there is potential for data loss when doing large scale
data removal.

2. they just made a change and are likely to want to immediately use
it - such as creating a new group. This will cause issues if they're
concurrently rebuilding tables and so might not have the change
available momentarily. E.g. the new group is suddenly removed while
doing an insert with it - bad UX.

Another way of looking at these points is: rebuilds should only happen
when we can do nothing else. So now we only rebuild on privacy changes
when it was not you authoring the change. This still allows users to
remove others and have their caches invalidated, but assumes that bad
actors aren't going to voluntairily remove themselves from groups.