[pk] Revert test changes: rebuilds have the same issue as resets
A rebuild can concurrently interact with a local write. If it orders >
by lww rules it will revert it. This is an inherent flaw with the
design that vector clocks will mitigate.

Rebuilds do help here compared to resets in that they greatly reduce
the window where this can happen. But the model doesn't capture this
and still fails.